requirements.txt
aiogram==3.15.0
openai>=1.59.2
aiosqlite==0.20.0
APScheduler==3.10.4
yookassa==3.1.0
python-dotenv==1.0.1
aiohttp>=3.9.0,<3.11
